moved auth.ts into backend in the lib directory, created auth-client in the frontend lib/aith-client.ts
