+++ backend/lib/auth.ts
@@ -1,6 +1,6 @@
 import { betterAuth } from "better-auth";
 import { prismaAdapter } from "better-auth/adapters/prisma";
-import prismaClientInstance from "../backend/prismaClient"
+import prismaClientInstance from "../prismaClient"
 export const auth = betterAuth({
   database: prismaAdapter(prismaClientInstance, {
         provider: 'mysql'
