+++ backend/server.ts
@@ -1,7 +1,7 @@
 //import express and dotenv
 import express from 'express';
 import { toNodeHandler } from "better-auth/node";
-import { auth } from "../utils/auth";
+import { auth } from "./lib/auth";
 import dotenv from 'dotenv';
 dotenv.config();
